+++ tket/src/OpType/OpTypeFunctions.cpp
@@ -108,7 +108,7 @@ const OpTypeSet& all_classical_types() {
       OpType::CopyBits,           OpType::RangePredicate,
       OpType::ExplicitPredicate,  OpType::ExplicitModifier,
       OpType::MultiBit,           OpType::WASM,
-      OpType::ClassicalExpBox,
+      OpType::ClassicalExpBox,    OpType::ClExpr,
   };
   static std::unique_ptr<const OpTypeSet> gates =
       std::make_unique<const OpTypeSet>(optypes);